refactor: order, order-table 병합
