+++ src/test/java/kitchenpos/order/application/OrderServiceTest.java
@@ -2,7 +2,7 @@ package kitchenpos.order.application;
 
 import kitchenpos.infra.client.FakeKitchenridersClient;
 import kitchenpos.order.domain.InMemoryOrderRepository;
-import kitchenpos.ordertable.domain.InMemoryOrderTableRepository;
+import kitchenpos.order.domain.InMemoryOrderTableRepository;
 import kitchenpos.menu.domain.InMemoryMenuRepository;
 import kitchenpos.menu.domain.MenuRepository;
 import kitchenpos.order.domain.Order;
@@ -10,8 +10,8 @@ import kitchenpos.order.domain.OrderLineItem;
 import kitchenpos.order.domain.OrderRepository;
 import kitchenpos.order.domain.OrderStatus;
 import kitchenpos.order.domain.OrderType;
-import kitchenpos.ordertable.domain.OrderTable;
-import kitchenpos.ordertable.domain.OrderTableRepository;
+import kitchenpos.order.domain.OrderTable;
+import kitchenpos.order.domain.OrderTableRepository;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
