+++ src/test/java/kitchenpos/order/domain/InMemoryOrderRepository.java
@@ -1,7 +1,5 @@
 package kitchenpos.order.domain;
 
-import kitchenpos.ordertable.domain.OrderTable;
-
 import java.util.*;
 
 public class InMemoryOrderRepository implements OrderRepository {
