+++ src/test/java/kitchenpos/order/domain/InMemoryOrderTableRepository.java
@@ -1,7 +1,7 @@
-package kitchenpos.ordertable.domain;
+package kitchenpos.order.domain;
 
-import kitchenpos.ordertable.domain.OrderTable;
-import kitchenpos.ordertable.domain.OrderTableRepository;
+import kitchenpos.order.domain.OrderTable;
+import kitchenpos.order.domain.OrderTableRepository;
 
 import java.util.*;
